chore: move chart props, event type to types folder
