+++ apps/chart/src/charts/lineChart.ts
@@ -1,4 +1,4 @@
-import Chart, { SelectSeriesInfo } from './chart';
+import Chart from './chart';
 
 import dataRange from '@src/store/dataRange';
 import scale from '@src/store/scale';
@@ -38,12 +38,7 @@ import {
   PlotLine,
   PlotBand,
 } from '@t/options';
-
-export interface LineChartProps {
-  el: HTMLElement;
-  options: LineChartOptions;
-  data: LineSeriesData;
-}
+import { LineChartProps, SelectSeriesInfo } from '@t/charts';
 
 /**
  * @class
