+++ apps/chart/src/charts/lineScatterChart.ts
@@ -1,4 +1,4 @@
-import Chart, { AddSeriesDataInfo, SelectSeriesInfo } from './chart';
+import Chart from './chart';
 
 import dataRange from '@src/store/dataRange';
 import scale from '@src/store/scale';
@@ -37,12 +37,7 @@ import {
   ScatterSeriesInput,
 } from '@t/options';
 import { RawSeries } from '@t/store/store';
-
-export interface LineScatterChartProps {
-  el: HTMLElement;
-  options: LineScatterChartOptions;
-  data: LineScatterData;
-}
+import { LineScatterChartProps, AddSeriesDataInfo, SelectSeriesInfo } from '@t/charts';
 
 /**
  * @class
