+++ apps/chart/src/component/areaSeries.ts
@@ -47,7 +47,7 @@ import { getValueAxisName } from '@src/helpers/axes';
 import { getDataLabelsOptions } from '@src/helpers/dataLabels';
 import { PointDataLabel } from '@t/components/dataLabels';
 import { AreaChartSeriesTheme, DotTheme } from '@t/theme';
-import { SelectSeriesInfo } from '@src/charts/chart';
+import { SelectSeriesInfo } from '@t/charts';
 import { message } from '@src/message';
 import { isAvailableSelectSeries, isAvailableShowTooltipInfo } from '@src/helpers/validation';
