+++ apps/chart/src/component/lineSeries.ts
@@ -39,7 +39,7 @@ import { getValueAxisName } from '@src/helpers/axes';
 import { getDataLabelsOptions } from '@src/helpers/dataLabels';
 import { PointDataLabel } from '@t/components/dataLabels';
 import { DotTheme, LineChartSeriesTheme } from '@t/theme';
-import { SelectSeriesInfo } from '@src/charts/chart';
+import { SelectSeriesInfo } from '@t/charts';
 import { message } from '@src/message';
 import { isAvailableSelectSeries, isAvailableShowTooltipInfo } from '@src/helpers/validation';
