+++ apps/chart/types/components/series.d.ts
@@ -5,17 +5,11 @@ import { LineModel, LabelModel } from './axis';
 import { SectorStyle, SectorStyleName } from '../../src/brushes/sector';
 import { LegendData } from './legend';
 import { TreemapSeriesData } from '../store/store';
-import Component from '../../src/component/component';
 import { BubbleDataLabel } from '../theme';
 
 export type Nullable<T> = T | null;
 export type StyleProp<T, K> = (T | K)[];
 export type PointModel = Point & { value?: number; name?: string };
-
-export type RespondersModel = {
-  component: Component;
-  detected: ResponderModel[];
-}[];
 
 export interface CircleStyle {
   strokeStyle?: string;
