+++ apps/chart/types/eventEmitter.d.ts
@@ -1,0 +1,11 @@
+export type CustomEventType =
+  | 'clickLegendLabel'
+  | 'clickLegendCheckbox'
+  | 'selectSeries'
+  | 'unselectSeries'
+  | 'hoverSeries'
+  | 'unhoverSeries'
+  | 'zoom'
+  | 'resetZoom';
+
+export type EventListener = (evt: any) => void;
